+++ pulse2percept/models/tests/test_beyeler2019.py
@@ -19,7 +19,7 @@ from pulse2percept.utils.testing import assert_warns_msg
 
 def test_ScoreboardSpatial():
     # ScoreboardSpatial automatically sets `rho`:
-    model = ScoreboardSpatial(engine='serial', xystep=5)
+    model = ScoreboardSpatial(xystep=5)
 
     # User can set `rho`:
     model.rho = 123
@@ -46,7 +46,7 @@ def test_ScoreboardSpatial():
     npt.assert_almost_equal(percept.data, 0)
 
     # Multiple frames are processed independently:
-    model = ScoreboardSpatial(engine='serial', rho=200, xystep=5,
+    model = ScoreboardSpatial(rho=200, xystep=5,
                               xrange=(-20, 20), yrange=(-15, 15))
     model.build()
     percept = model.predict_percept(ArgusI(stim={'A1': [1, 0], 'B3': [0, 2]}))
@@ -80,7 +80,7 @@ def test_deepcopy_ScoreboardSpatial():
 
 def test_ScoreboardModel():
     # ScoreboardModel automatically sets `rho`:
-    model = ScoreboardModel(engine='serial', xystep=5)
+    model = ScoreboardModel(xystep=5)
     npt.assert_equal(model.has_space, True)
     npt.assert_equal(model.has_time, False)
     npt.assert_equal(hasattr(model.spatial, 'rho'), True)
@@ -108,7 +108,7 @@ def test_ScoreboardModel():
     npt.assert_almost_equal(model.predict_percept(implant).data, 0)
 
     # Multiple frames are processed independently:
-    model = ScoreboardModel(engine='serial', rho=200, xystep=5,
+    model = ScoreboardModel(rho=200, xystep=5,
                             xrange=(-20, 20), yrange=(-15, 15))
     model.build()
     percept = model.predict_percept(ArgusI(stim={'A1': [1, 2]}))
@@ -156,14 +156,14 @@ def test_ScoreboardModel_predict_percept():
     npt.assert_almost_equal(percept.data[33, 46, 0], np.max(percept.data))
 
     # Full Argus II: 60 bright spots
-    model = ScoreboardModel(engine='serial', xystep=0.55, rho=100)
+    model = ScoreboardModel(xystep=0.55, rho=100)
     model.build()
     percept = model.predict_percept(ArgusII(stim=np.ones(60)))
     npt.assert_equal(np.sum(np.isclose(percept.data, 0.8, rtol=0.1, atol=0.1)),
                      88)
 
     # Model gives same outcome as Spatial:
-    spatial = ScoreboardSpatial(engine='serial', xystep=1, rho=100)
+    spatial = ScoreboardSpatial(xystep=1, rho=100)
     spatial.build()
     spatial_percept = model.predict_percept(ArgusII(stim=np.ones(60)))
     npt.assert_almost_equal(percept.data, spatial_percept.data)
@@ -176,10 +176,9 @@ def test_ScoreboardModel_predict_percept():
     assert_warns_msg(UserWarning, model.predict_percept, msg, implant)
 
 
-@ pytest.mark.parametrize('engine', ('serial', 'cython'))
-def test_AxonMapSpatial(engine):
+def test_AxonMapSpatial():
     # AxonMapSpatial automatically sets `rho`, `axlambda`:
-    model = AxonMapSpatial(engine=engine, xystep=5)
+    model = AxonMapSpatial(xystep=5)
 
     # User can set `rho`:
     model.rho = 123
@@ -211,7 +210,7 @@ def test_AxonMapSpatial(engine):
         AxonMapSpatial(axlambda=9).build()
 
     # Multiple frames are processed independently:
-    model = AxonMapSpatial(engine=engine, rho=200, axlambda=100, xystep=5,
+    model = AxonMapSpatial(rho=200, axlambda=100, xystep=5,
                            xrange=(-20, 20), yrange=(-15, 15))
     model.build()
     percept = model.predict_percept(ArgusI(stim={'A1': [1, 0], 'B3': [0, 2]}))
@@ -273,9 +272,8 @@ def test_AxonMapSpatial_plot():
         plt.close(fig)
 
 
-@ pytest.mark.parametrize('engine', ('serial', 'cython'))
-def test_AxonMapModel(engine):
-    set_params = {'xystep': 2, 'engine': engine, 'rho': 432, 'axlambda': 20,
+def test_AxonMapModel():
+    set_params = {'xystep': 2, 'rho': 432, 'axlambda': 20,
                   'n_axons': 9, 'n_ax_segments': 50,
                   'xrange': (-30, 30), 'yrange': (-20, 20),
                   'loc_od': (5, 6)}
@@ -351,11 +349,10 @@ def test_deepcopy_AxonMapModel(build):
 @ pytest.mark.parametrize('eye', ('LE', 'RE'))
 @ pytest.mark.parametrize('loc_od', ((15.5, 1.5), (7.0, 3.0), (-2.0, -2.0)))
 @ pytest.mark.parametrize('sign', (-1.0, 1.0))
-@ pytest.mark.parametrize('engine', ('serial', 'cython'))
-def test_AxonMapModel__jansonius2009(eye, loc_od, sign, engine):
+def test_AxonMapModel__jansonius2009(eye, loc_od, sign):
     # With `rho` starting at 0, all axons should originate in the optic disc
     # center
-    model = AxonMapModel(loc_od=loc_od, xystep=2, engine=engine,
+    model = AxonMapModel(loc_od=loc_od, xystep=2,
                          ax_segments_range=(0, 45),
                          n_ax_segments=100)
     for phi0 in [-135.0, 66.0, 128.0]:
@@ -365,7 +362,7 @@ def test_AxonMapModel__jansonius2009(eye, loc_od, sign, engine):
 
     # These axons should all end at the meridian
     for phi0 in [110.0, 135.0, 160.0]:
-        model = AxonMapModel(loc_od=(15, 2), xystep=2, engine=engine,
+        model = AxonMapModel(loc_od=(15, 2), xystep=2,
                              n_ax_segments=801,
                              ax_segments_range=(0, 45))
         ax_pos = model.spatial._jansonius2009(sign * phi0)
@@ -374,31 +371,28 @@ def test_AxonMapModel__jansonius2009(eye, loc_od, sign, engine):
     # `phi0` must be within [-180, 180]
     for phi0 in [-200.0, 181.0]:
         with pytest.raises(ValueError):
-            failed = AxonMapModel(xystep=2, engine=engine)
+            failed = AxonMapModel(xystep=2)
             failed.spatial._jansonius2009(phi0)
 
     # `n_rho` must be >= 1
     for n_rho in [-1, 0]:
         with pytest.raises(ValueError):
-            model = AxonMapModel(n_ax_segments=n_rho, xystep=2,
-                                 engine=engine)
+            model = AxonMapModel(n_ax_segments=n_rho, xystep=2)
             model.spatial._jansonius2009(0.0)
 
     # `ax_segments_range` must have min <= max
     for lorho in [-200.0, 90.0]:
         with pytest.raises(ValueError):
-            model = AxonMapModel(ax_segments_range=(lorho, 45), xystep=2,
-                                 engine=engine)
+            model = AxonMapModel(ax_segments_range=(lorho, 45), xystep=2)
             model.spatial._jansonius2009(0)
     for hirho in [-200.0, 40.0]:
         with pytest.raises(ValueError):
-            model = AxonMapModel(ax_segments_range=(45, hirho), xystep=2,
-                                 engine=engine)
+            model = AxonMapModel(ax_segments_range=(45, hirho), xystep=2)
             model.spatial._jansonius2009(0)
 
     # A single axon fiber with `phi0`=0 should return a single pixel location
     # that corresponds to the optic disc
-        model = AxonMapModel(loc_od=loc_od, xystep=2, engine=engine, eye=eye,
+        model = AxonMapModel(loc_od=loc_od, xystep=2, eye=eye,
                              ax_segments_range=(0, 0),
                              n_ax_segments=1)
         single_fiber = model.spatial._jansonius2009(0)
@@ -406,19 +400,17 @@ def test_AxonMapModel__jansonius2009(eye, loc_od, sign, engine):
         npt.assert_almost_equal(single_fiber[0], loc_od)
 
 
-@ pytest.mark.parametrize('engine', ('serial', 'cython'))
-def test_AxonMapModel_grow_axon_bundles(engine):
+def test_AxonMapModel_grow_axon_bundles():
     for n_axons in [1, 2, 3, 5, 10]:
-        model = AxonMapModel(xystep=2, engine=engine, n_axons=n_axons,
+        model = AxonMapModel(xystep=2, n_axons=n_axons,
                              axons_range=(-20, 20), xrange=(-20, 20),
                              yrange=(-15, 15))
         bundles = model.spatial.grow_axon_bundles()
         npt.assert_equal(len(bundles), n_axons)
 
 
-@ pytest.mark.parametrize('engine', ('serial', 'cython'))
-def test_AxonMapModel_find_closest_axon(engine):
-    model = AxonMapModel(xystep=1, engine=engine, n_axons=5,
+def test_AxonMapModel_find_closest_axon():
+    model = AxonMapModel(xystep=1, n_axons=5,
                          xrange=(-20, 20), yrange=(-15, 15),
                          axons_range=(-45, 45))
     model.build()
@@ -448,9 +440,8 @@ def test_AxonMapModel_find_closest_axon(engine):
     npt.assert_equal(closest_idx, 0)
 
 
-@ pytest.mark.parametrize('engine', ('serial', 'cython'))
-def test_AxonMapModel_calc_axon_sensitivity(engine):
-    model = AxonMapModel(xystep=2, engine=engine, n_axons=10,
+def test_AxonMapModel_calc_axon_sensitivity():
+    model = AxonMapModel(xystep=2, n_axons=10,
                          xrange=(-20, 20), yrange=(-15, 15),
                          axons_range=(-30, 30))
     model.build()
@@ -493,9 +484,8 @@ def test_AxonMapModel_calc_axon_sensitivity_deprecated_pad(pad):
         model.spatial.calc_axon_sensitivity(axons)
 
 
-@ pytest.mark.parametrize('engine', ('serial', 'cython'))
-def test_AxonMapModel_calc_bundle_tangent(engine):
-    model = AxonMapModel(xystep=5, engine=engine, n_axons=500,
+def test_AxonMapModel_calc_bundle_tangent():
+    model = AxonMapModel(xystep=5, n_axons=500,
                          xrange=(-20, 20), yrange=(-15, 15),
                          n_ax_segments=500, axons_range=(-180, 180),
                          ax_segments_range=(3, 50))
@@ -509,9 +499,8 @@ def test_AxonMapModel_calc_bundle_tangent(engine):
         model.spatial.calc_bundle_tangent(0, [1000])
 
 
-@ pytest.mark.parametrize('engine', ('serial', 'cython'))
-def test_AxonMapModel_calc_bundle_tangent_fast(engine):
-    model = AxonMapModel(xystep=5, engine=engine, n_axons=500,
+def test_AxonMapModel_calc_bundle_tangent_fast():
+    model = AxonMapModel(xystep=5, n_axons=500,
                          xrange=(-20, 20), yrange=(-15, 15),
                          n_ax_segments=500, axons_range=(-180, 180),
                          ax_segments_range=(3, 50))
@@ -525,10 +514,9 @@ def test_AxonMapModel_calc_bundle_tangent_fast(engine):
 
 
 
-@ pytest.mark.parametrize('engine', ('serial', 'cython'))
-def test_AxonMapModel_predict_percept(engine):
+def test_AxonMapModel_predict_percept():
     model = AxonMapModel(xystep=0.55, axlambda=100, rho=100,
-                         thresh_percept=0, engine=engine,
+                         thresh_percept=0,
                          xrange=(-20, 20), yrange=(-15, 15),
                          n_axons=500)
     model.build()
@@ -551,7 +539,7 @@ def test_AxonMapModel_predict_percept(engine):
     npt.assert_almost_equal(np.sum(percept.data[39:, :, 0]), 0)
 
     # Full Argus II with small lambda: 60 bright spots
-    model = AxonMapModel(engine='serial', xystep=1, rho=100, axlambda=40,
+    model = AxonMapModel(xystep=1, rho=100, axlambda=40,
                          xrange=(-20, 20), yrange=(-15, 15), n_axons=500)
     model.build()
     percept = model.predict_percept(ArgusII(stim=np.ones(60)))
@@ -561,7 +549,7 @@ def test_AxonMapModel_predict_percept(engine):
     npt.assert_equal(np.sum(percept.data > 0.275), 56)
 
     # Model gives same outcome as Spatial:
-    spatial = AxonMapSpatial(engine='serial', xystep=1, rho=100, axlambda=40,
+    spatial = AxonMapSpatial(xystep=1, rho=100, axlambda=40,
                              xrange=(-20, 20), yrange=(-15, 15), n_axons=500)
     spatial.build()
     spatial_percept = spatial.predict_percept(ArgusII(stim=np.ones(60)))